Agrego avatar, banner del perfil, ahora se puede editar y eliminar publicaciones

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Validator;
 use App\Models\Post;
 use App\Http\Requests\Validaciones;
+use App\Http\Requests\ValidacionesUpdate;
 use DateTime;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Crypt;
@@ -23,9 +24,9 @@ class PostController extends Controller
     {
         $content = $request->content;
         $created_at = new DateTime();
+        $validations = new Validaciones();
         $updated_at = $created_at;
 
-        $validations = new Validaciones();
 
         $validator = Validator::make($request->all(), $validations->rules(), $validations->messages());
         if($validator->fails()){
@@ -42,33 +43,41 @@ class PostController extends Controller
         
     }
 
-    public function update( Request $valores)
+    public function update( Request $request)
     {
-        $contenido = $valores->contenido;
-        $fecha_publicacion = $valores->fecha_publicacion;
-        $fecha_actualizacion = $valores->fecha_actualizacion;
+        $validations = new ValidacionesUpdate();
+        $idPost = Crypt::decrypt($request->id);
+        $content = "contentEdit".$idPost;
+        $content = $request->$content;
+        $created_at = new DateTime();
+        $updated_at = $created_at;
+        $validator = Validator::make($request->all(), $validations->rules($idPost), $validations->messages());
 
-        $validaciones = new Validaciones();
-
-        $validator = Validator::make($valores->all(), $validaciones->rules(), $validaciones->messages());
         if($validator->fails()){
-            return redirect('admin')->with('status', 'No puede enviar datos vacios!');
+            return redirect()->back()->withErrors($validator)->withInput();
         }
-        $post = Post::find(Crypt::decrypt($valores->id));
-        $post->contenido = $contenido;
-        $post->autor = auth()->user()->name;
-        $post->created_at = $fecha_publicacion;
-        $post->updated_at = $fecha_actualizacion;
-        $post->save();
-
-        return redirect('admin')->with('status', 'Post actualizado exitosamente!');
-
+        $post = Post::find($idPost);
+        if($post->author == Auth::user()->name){
+            $post->content = $content;
+            $post->created_at = $created_at;
+            $post->updated_at = $updated_at;
+            $post->save();
+        }else{
+            return redirect()->back()->withErrors("Error desconocido");
+        }
+        sleep(1);
+        return redirect()->back();
     }
 
-    public function delete($id)
+    public function delete( Request $request)
     {
-        $post = Post::find(Crypt::decrypt($id));
-        $post->delete();
+        $idPost = Crypt::decrypt($request->idPost);
+        $post = Post::find($idPost);
+        if($post->author == Auth::user()->name){
+            $post->delete();
+        }else{
+            return redirect()->back()->withErrors("Error desconocido");
+        }
         sleep(1);
         return redirect()->back();
